+++ .gitignore
@@ -6,3 +6,4 @@
 /out
 /src/META-INF
 /test.lesson2
+/src/lesson7/temp/chat-app/
