--- src/lesson6/chat_app/network/src/chat/network/ServerSocketThreadListener.java
@@ -1,20 +1,0 @@
-package chat.network;
-
-import java.net.ServerSocket;
-import java.net.Socket;
-
-public interface ServerSocketThreadListener {
-    void onServerStart(ServerSocketThread thread);
-
-    void onServerStop(ServerSocketThread thread);
-
-    void onServerSocketCreated(ServerSocketThread thread, ServerSocket server);
-
-    void onServerTimeout(ServerSocketThread thread, ServerSocket server);
-
-    void onSocketAccepted(ServerSocketThread thread, ServerSocket server, Socket socket);
-
-    void onServerException(ServerSocketThread thread, Throwable exception);
-
-    void broadcastMessage(String msg);
-}
