--- src/lesson6/chat_app/network/src/chat/network/SocketThreadListener.java
@@ -1,16 +1,0 @@
-package chat.network;
-
-import java.net.Socket;
-
-public interface SocketThreadListener {
-
-    void onSocketStart(SocketThread thread, Socket socket);
-
-    void onSocketStop(SocketThread thread);
-
-    void onSocketReady(SocketThread thread, Socket socket);
-
-    void onReceiveString(SocketThread thread, Socket socket, String msg);
-
-    void onSocketException(SocketThread thread, Exception exception);
-}
