+++ src/lesson6/qq.txt
@@ -1,0 +1,2 @@
+Вопросы к 6 уроку:
+Не могу разобраться как работает таймаут в классе ServerSocketThread, куда передается эта переменная, какой метод ее использует, что прерывает accept?
