+++ src/lesson7/chat_app/chat-server/src/chat/server/core/ChatServerListener.java
@@ -1,0 +1,5 @@
+package chat.server.core;
+
+public interface ChatServerListener {
+    void onChatServerMessage(String msg);
+}
