+++ src/lesson7/chat_app/network/src/chat/network/ServerSocketThread.java
@@ -1,0 +1,44 @@
+package chat.network;
+
+import java.io.IOException;
+import java.net.ServerSocket;
+import java.net.Socket;
+import java.net.SocketTimeoutException;
+
+public class ServerSocketThread extends Thread {
+    private final int port;
+    private final int timeout;
+    private final ServerSocketThreadListener listener;
+
+    public ServerSocketThread(ServerSocketThreadListener listener, String name, int port, int timeout) {
+        super(name);
+        this.port = port;
+        this.timeout = timeout;
+        this.listener = listener;
+        start();
+    }
+
+    @Override
+    public void run() {
+
+        listener.onServerStart(this);
+        try (ServerSocket server = new ServerSocket(port)) {
+            server.setSoTimeout(timeout);
+            listener.onServerSocketCreated(this, server);
+            while (!isInterrupted()) { //бесконечный цикл создающий сокеты
+                  Socket client;
+                try {
+                    client = server.accept();
+                } catch (SocketTimeoutException e) {
+                    listener.onServerTimeout(this, server);
+                    continue;
+                }
+                listener.onSocketAccepted(this, server, client);
+            }
+        } catch (IOException e) {
+            listener.onServerException(this, e);
+        } finally {
+            listener.onServerStop(this);
+        }
+    }
+}
